attempt displaying

+++ app/src/main/java/com/bfr/opencvapp/utils/TfLiteMidasMultiOut.java
@@ -135,7 +135,7 @@ public class TfLiteMidasMultiOut {
      * @param bitmap original image in bitmap format
      * @return array of detections
      */
-    public float[] recognizeImage(Bitmap bitmap) {
+    public int[] recognizeImage(Bitmap bitmap) {
 
         ByteBuffer byteBuffer = convertBitmapToByteBuffer(bitmap);
 
@@ -156,7 +156,7 @@ public class TfLiteMidasMultiOut {
 
 
 
-        float[] result = new float[embeedings[0].length * embeedings[0][0].length];
+        int[] result = new int[embeedings[0].length * embeedings[0][0].length];
         Log.i("coucou", "i=" + embeedings[0].length + " j=" + embeedings[0][0].length);
         for (int i = 0; i < embeedings[0].length; i++) {
             for (int j = 0; j < embeedings[0][i].length; j++) {
